# services/data-service/src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Escape Doom Data API
  version: 0.0.1
servers:
  - url: https://api.server.test/v1/
paths:
...
  /template/create:
    post:
      operationId: createTemplate
      tags:
        - Template
      summary: Creates a new Template for Escape Doom Game
      description: Creates a new Template for EscapeRoom
      requestBody:
        description: Lectors ID + Name and Description for a Template
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EscapeRoomTemplateCreateRequestDTO'
      responses:
        '200':
          $ref: '#/components/responses/EscapeRoomTemplateResultDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /template/delete/{escape-room-template-id}:
    delete:
      operationId: deleteTemplate
      tags:
        - Template
      summary: Deletes an EscapeRoomTemplate
      description: Deletes an EscapeRoomTemplate by its unique ID
      parameters:
        - name: escape-room-template-id
          in: path
          required: true
          description: The unique ID of the EscapeRoomTemplate to delete
          schema:
            type: string
            example: b6557071-e7fa-47bc-bdd1-5657ebd325b8
      responses:
        '200':
          $ref: '#/components/responses/EscapeRoomTemplateResultDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /template/override/{escape-room-template-id}:
    put:
      operationId: putTemplate
      tags:
        - Template
      summary: Overrides an existing EscapeRoomTemplate
      description: >-
        Override the name, description, and levels of an existing
        EscapeRoomTemplate
      parameters:
        - name: escape-room-template-id
          in: path
          required: true
          description: The unique ID of the EscapeRoomTemplate to update
          schema:
            type: string
            example: b6557071-e7fa-47bc-bdd1-5657ebd325b8
      requestBody:
        description: The updated data for the template
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EscapeRoomTemplateUpdateRequestDTO'
      responses:
        '200':
          description: Template updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EscapeRoomTemplateUpdateResultDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /all-templates:
    get:
      operationId: getAllTemplates
      tags:
        - Template
      summary: Get all EscapeRoomTemplates
      description: Retrieve a list of all existing EscapeRoomTemplates from a Lector
      responses:
        '200':
          description: A list of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EscapeRoomTemplateDTO'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /template/{escape-room-template-id}:
    get:
      operationId: getTemplate
      tags:
        - Template
      summary: Get a specific EscapeRoomTemplate by ID
      description: Retrieve details of a specific EscapeRoomTemplate using its unique ID
      parameters:
        - name: escape-room-template-id
          in: path
          required: true
          description: The unique ID of the EscapeRoomTemplate
          schema:
            type: string
            example: b6557071-e7fa-47bc-bdd1-5657ebd325b8
      responses:
        '200':
          description: Details of the specified template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EscapeRoomTemplateDTO'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                type: object
                title: GetTemplateNotFoundDTO
                properties:
                  message:
                    type: string
                    example: "EscapeRoomTemplate with the provided ID was not found"
                  code:
                    type: number
                    example: 404
        '500':
          $ref: '#/components/responses/InternalServerError'


  ### Level
...
components:
  schemas:
    NodeType:
      description: The type of a node
      type: string
      enum:
        - ZOOM
        - CONSOLE
        - DETAIL
        - STORY
      example: "ZOOM"
    NodeDTO:
      description: A node of an escape-room instance
      type: object
      properties:
        node_type:
          $ref: '#/components/schemas/NodeType'
        node_info:
          $ref: '#/components/schemas/NodeInfoDTO'
        position:
          $ref: '#/components/schemas/PositionDTO'
    NodeInfoDTO:
      description: The information of a node
      type: object
      properties:
        description:
          type: string
          description: The description of the node
          example: This is a story node
        title:
          type: string
          description: The display title of the node
          example: I like cheese
        imageURI:
          type: string
          description: The URI of the image to display
          example: https://example.com/image.png
    PositionDTO:
      description: The position of a node
      type: object
      properties:
        top:
          type: number
          format: double
          description: The relative top position of the node
          example: 50.5
        left:
          type: number
          format: double
          description: The relative left position of the node
          example: 22


    # Enums
    CodingLanguage:
      type: string
      enum:
        - JAVA
        - TYPESCRIPT
      example: "JAVA"
    # Classes
    EscapeRoomTemplateDTO:
      description: The escape-room template, base for an EscapeRoom Game
      type: object
      properties:
        escape_room_template_id:
          type: string
          description: Unique ID for the Template
          example: b6557071-e7fa-47bc-bdd1-5657ebd325b8
        name:
          type: string
          description: Name of the Template
          example: SDE24
        description:
          type: string
          description: Description of the Template
          example: Cäsar´s Rätsel
    EscapeRoomTemplateCreateRequestDTO:
      description: The escape-room template, base for an EscapeRoom Game
      type: object
      properties:
        name:
          type: string
          description: Name of the Template
          example: SDE24
        description:
          type: string
          description: Description of the Template
          example: Cäsar´s Rätsel
    EscapeRoomTemplateResultDTO:
      description: The result of an operation on an EscapeRoomTemplate
      type: object
      properties:
        message:
          type: string
          description: >-
            A message indicating the result of the operation (e.g., creation or
            deletion)
          example: Operation successful
    EscapeRoomTemplateUpdateRequestDTO:
      description: The payload for updating an EscapeRoomTemplate
      type: object
      properties:
        name:
          type: string
          description: The updated name of the template
          example: SDE25
        description:
          type: string
          description: The updated description of the template
          example: Updated description for Caesar's Riddle
        levels:
          type: array
          items:
            $ref: '#/components/schemas/EscapeRoomLevelDTO'
          description: The updated levels for the escape-room template
    EscapeRoomTemplateUpdateResultDTO:
      description: The result of updating an EscapeRoomTemplate
      type: object
      properties:
        message:
          type: string
          description: A message indicating the result of the update operation
          example: Template updated successfully
        code:
          type: number
          description: A status code representing the operation result
          example: 200
    EscapeRoomLevelDTO:
      description: The current level of an escape-room instance
      type: object
      properties:
        escape_room_template_id:
          type: string
          description: The unique ID of the escape room template
          example: "d1087efe-41fd-42da-9110-62d35659cf1f"
        escape_room_level_id:
          type: string
          description: The unique ID of the escape room level
          example: "d1087efe-41fd-42da-9110-62d35659cf1f"
        sequence:
          type: integer
          description: The sequence number of the level
          example: 1
        scenes:
          type: array
          items:
            $ref: "#/components/schemas/SceneDTO"
          description: List of scenes in the level
        riddles:
          type: array
          items:
            $ref: "#/components/schemas/RiddleDTO"
          description: List of riddles in the level
    SceneDTO:
      description: A scene of an escape-room instance
      type: object
      properties:
        escape_room_level_id:
          type: string
          description: The ID of the escape room level that contains this riddle
          example: "a12b34c5-6789-4def-abcd-12345678abcd"
        escape_room_sequence_id:
          type: string
          description: The unique ID of the scene sequence
          example: "241a70fe-47d6-4756-9ac7-330f1b199e84"
        scene_sequence:
          type: integer
          description: Defines the "position" of the scene in the escape room
          example: 1
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/NodeDTO"
          description: The nodes of the scene
        background_image_uri:
          type: string
          description: The URI of the background image
          example: "https://example.com/background.png"
        name:
          type: string
          description: The name of the scene
          example: "Scene 1"
    RiddleDTO:
      description: Base schema for a riddle
      type: object
      properties:
        escape_room_riddle_id:
          type: string
          description: The unique ID of the riddle
          example: "5830daed-cb7f-47dd-8248-5dee9bf0aa3d"
        escape_room_level_id:
          type: string
          description: The ID of the escape room level that contains this riddle
          example: "a12b34c5-6789-4def-abcd-12345678abcd"
        language:
          $ref: "#/components/schemas/CodingLanguage"
        function_signature:
          type: string
          description: The function signature
          example: "public static int sum(int a, int b)"
        input:
          type: string
          description: The input values for the function
          example: "2, 3"
        variable_name:
          type: string
          description: The name of the variable to compare
          example: "result"
        expected_output:
          type: string
          description: The expected output of the riddle
          example: "42"
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            title: CreateBadRequestDTO
            properties:
              message:
                type: string
                example: Invalid data provided
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            title: CreateNotFoundDTO
            properties:
              message:
                type: string
                example: The requested resource was not found
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            title: CreateInternalServerErrorDTO
            properties:
              message:
                type: string
                example: An unexpected error occurred on the server
    EscapeRoomTemplateResultDTO:
      description: Operation result for EscapeRoomTemplate
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EscapeRoomTemplateResultDTO'